+++ src/RadixDlt.NetworkGateway.PostgresIntegration/ServiceCollectionExtensions.cs
@@ -104,6 +104,7 @@ public static class ServiceCollectionExtensions
                 dataSourceBuilder.MapEnum<EntityType>();
                 dataSourceBuilder.MapEnum<LedgerTransactionStatus>();
                 dataSourceBuilder.MapEnum<LedgerTransactionType>();
+                dataSourceBuilder.MapEnum<LedgerTransactionManifestClass>();
                 dataSourceBuilder.MapEnum<LedgerTransactionMarkerType>();
                 dataSourceBuilder.MapEnum<LedgerTransactionMarkerEventType>();
                 dataSourceBuilder.MapEnum<LedgerTransactionMarkerOperationType>();
@@ -117,7 +118,6 @@ public static class ServiceCollectionExtensions
                 dataSourceBuilder.MapEnum<ModuleId>();
                 dataSourceBuilder.MapEnum<SborTypeKind>();
                 dataSourceBuilder.MapEnum<StateType>();
-                dataSourceBuilder.MapEnum<LedgerTransactionMarkerManifestClass>();
 
                 return new NpgsqlDataSourceHolder<T>(dataSourceBuilder.Build());
             },